chore: move Flex and Grid to f36-core
