--- packages/components/layout/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@contentful/f36-layout",
-  "version": "4.0.0",
-  "private": true,
-  "description": "Forma 36: Layout React Components",
-  "license": "MIT",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "main": "dist/main.js",
-  "module": "dist/module.js",
-  "types": "dist/types.d.ts",
-  "source": "src/index.ts",
-  "sideEffects": false,
-  "browserslist": "extends @contentful/browserslist-config",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/contentful/forma-36"
-  },
-  "scripts": {
-    "build": "parcel build index.ts"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.6.2",
-    "emotion": "^10.0.17",
-    "@contentful/f36-tokens": "^4.0.0",
-    "@contentful/f36-core": "^4.0.0",
-    "csstype": "^3.0.5"
-  },
-  "peerDependencies": {
-    "react": ">=16.8"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
