--- packages/components/layout/src/index.ts
@@ -1,2 +1,0 @@
-export * from './Grid';
-export * from './Flex';
