+++ packages/forma-36-react-components/package.json
@@ -33,7 +33,6 @@
     "@babel/runtime": "^7.3.4",
     "@contentful/f36-badge": "^4.0.0",
     "@contentful/f36-core": "^4.0.0",
-    "@contentful/f36-layout": "^4.0.0",
     "@contentful/f36-tokens": "^4.0.0",
     "@contentful/forma-36-fcss": "^4.0.0",
     "@contentful/f36-utils": "^4.0.0",
