+++ packages/forma-36-react-components/src/components/Asset/Asset.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { Meta, Story } from '@storybook/react/types-6-0';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 
 import { Asset, types } from './Asset';
 import { Paragraph, SectionHeading } from '../Typography';
