+++ packages/forma-36-react-components/src/components/Card/Card.stories.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import type { Meta, Story } from '@storybook/react/types-6-0';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 
 import { Heading, Paragraph, SectionHeading, Typography } from '../Typography';
 import { Card, CardProps } from './Card';
