+++ packages/forma-36-react-components/src/components/Card/EntryCard/EntryCard.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 
 import { EntryCard, EntryCardProps } from './EntryCard';
 import { CardDragHandle } from '../CardDragHandle/CardDragHandle';
