+++ packages/forma-36-react-components/src/components/CheckboxField/CheckboxField.stories.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 
 import { FieldGroup } from '../Form';
 import { CheckboxField, CheckboxFieldProps } from './CheckboxField';
