+++ packages/forma-36-react-components/src/components/Notification/Notification.stories.tsx
@@ -4,7 +4,7 @@ import { Button } from '../Button';
 import { Notification, NotificationsAPI } from './Notification';
 import notes from './README.mdx';
 import { NotificationItem } from './NotificationItem';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 import { SectionHeading } from '../Typography';
 
 let index = 0;
