+++ packages/forma-36-react-components/src/components/RadioButton/RadioButton.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { RadioButton, RadioButtonProps } from './RadioButton';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 import { SectionHeading } from '../Typography';
 
 export default {
