+++ packages/forma-36-react-components/src/components/Switch/Switch.stories.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import { Switch, SwitchProps } from './Switch';
 import { SectionHeading } from '../Typography';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 
 export default {
   title: 'Components/Switch',
