+++ packages/forma-36-react-components/src/components/Tabs/Tabs.stories.tsx
@@ -6,7 +6,7 @@ import { Tabs } from './Tabs';
 import { Tab } from './Tab';
 import { TabPanel } from './TabPanel';
 import { SectionHeading } from '../Typography';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 
 export default {
   argTypes: {
