+++ packages/forma-36-react-components/src/components/TextLink/TextLink.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { TextLink, textLinkColor } from './TextLink';
 import { Paragraph } from '../Typography';
 import tokens from '@contentful/f36-tokens';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 import '@contentful/forma-36-fcss/dist/styles.css';
 
 import notes from './README.mdx';
