+++ packages/forma-36-react-components/src/components/Tooltip/Tooltip.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Tooltip } from './Tooltip';
 import { Paragraph, SectionHeading } from '../Typography';
 import { TextLink } from '../TextLink';
-import { Flex } from '@contentful/f36-layout';
+import { Flex } from '@contentful/f36-core';
 import notes from './README.mdx';
 
 export default {
